comment filtering

+++ admin/functions.php
@@ -123,15 +123,13 @@ function display_posts()
         echo "<td><img class='img-fluid' src='../images/posts/{$post_image}' alt='' style='max-width:100px;'/></td>";
         echo "<td>{$post_tags}</td>";
 
-        //$comment_count_query = "SELECT * FROM comments WHERE comment_post_id = " . $post_id;
-        //echo $post_id;
-        //$comment_query = mysqli_query($connection, $comment_count_query);
-        //$row = mysqli_fetch_array($comment_query);
-        //$comment_id = $row['comment_id'];
-        //sdecho $comment_id;
-        //$comment_count = mysqli_num_rows($comment_query);
+        $comment_count_query = "SELECT * FROM comments WHERE comment_post_id = " . $post_id;
+        $comment_query = mysqli_query($connection, $comment_count_query);
+        $row = mysqli_fetch_array($comment_query);
+        $comment_id = isset($row['comment_id']);
+        $comment_count = mysqli_num_rows($comment_query);
 
-        //echo "<td><a href='comment.php?id=$comment_id'>{$comment_count}</a></td>";
+        echo "<td><a href='comments-filtered.php?id=$comment_id'>{$comment_count}</a></td>";
 
         echo "<td>{$post_views_count}</td>";
         echo '<td class="text-right" style="min-width: 130px;">';
@@ -476,6 +474,50 @@ function display_comments()
         echo '</tr>';
     }
 }
+function display_comments_filtered()
+{
+    global $connection;
+    $query = "SELECT * FROM comments WHERE comment_post_id =" . mysqli_real_escape_string($connection, $_GET['id']);
+    $display_all_posts = mysqli_query($connection, $query);
+    while ($row = mysqli_fetch_assoc($display_all_posts)) {
+        $comment_id = $row['comment_id'];
+        $comment_post_id = $row['comment_post_id'];
+        $comment_date = $row['comment_date'];
+        $comment_author = $row['comment_author'];
+        $comment_email = $row['comment_email'];
+        $comment_content = $row['comment_content'];
+        $comment_status = $row['comment_status'];
+
+        echo '<tr>';
+        echo "<td>{$comment_id}</td>";
+        echo "<td>{$comment_author}</td>";
+        echo "<td>{$comment_content}</td>";
+        echo "<td>{$comment_email}</td>";
+        echo "<td>{$comment_status}</td>";
+
+        $query = "SELECT * FROM posts WHERE post_id = " . $comment_post_id;
+        $show_post_name = mysqli_query($connection, $query);
+        while ($row = mysqli_fetch_assoc($show_post_name)) {
+            $post_id = $row['post_id'];
+            $post_title = $row['post_title'];
+        }
+
+        echo "<td><a href='../post.php?p_id=$post_id'>{$post_title}</a></td>";
+        echo "<td>{$comment_date}</td>";
+        echo '<td class="text-right" style="min-width: 130px">';
+        echo '<a href ="comments.php?approve=' . $comment_id . '"
+        class="btn btn-success btn-sm btn-round btn-icon mr-2">
+        <i class="fal fa-thumbs-up pt-2"></i></a>';
+        echo '<a href ="comments.php?unapprove=' . $comment_id . '"
+        class="btn btn-warning btn-sm btn-round btn-icon mr-2">
+        <i class="fal fa-thumbs-down pt-2"></i></a>';
+        echo '<a onClick="javascript: return confirm(\'Are you sure you want to delete?\');" href ="comments.php?delete=' . $comment_id . '"
+        class="btn btn-danger btn-sm btn-round btn-icon">
+        <i class="fal fa-trash-alt pt-2"></i></a>';
+        echo '</td>';
+        echo '</tr>';
+    }
+}
 
 function approve_comment()
 {
@@ -610,12 +652,16 @@ function delete_user()
 {
     global $connection;
     if (isset($_GET['delete'])) {
-        $the_user_id = $_GET['delete'];
-        $query = 'DELETE FROM users WHERE user_id = ' . $the_user_id;
-        $delete_user = mysqli_query($connection, $query);
-        header("Location: users.php");
-        if (!$delete_user) {
-            die('Query Failed' . mysqli_error($connection));
+        if (isset($_SESSION['user_role'])) {
+            if ($_SESSION['user_role'] == 'Administrator') {
+                $the_user_id = $_GET['delete'];
+                $query = 'DELETE FROM users WHERE user_id = ' . $the_user_id;
+                $delete_user = mysqli_query($connection, $query);
+                header("Location: users.php");
+                if (!$delete_user) {
+                    die('Query Failed' . mysqli_error($connection));
+                }
+            }
         }
     }
 }
